tasks/requirements.txt
django
djangorestframework
drf-spectacular
djangorestframework-simplejwt
django-filter